[tuya] fix initialization bug (#320)

+++ bundles/org.smarthomej.binding.tuya/src/main/java/org/smarthomej/binding/tuya/internal/cloud/dto/DeviceSchema.java
@@ -51,8 +51,8 @@ public class DeviceSchema {
     }
 
     public static class NumericRange {
-        public int min = Integer.MIN_VALUE;
-        public int max = Integer.MAX_VALUE;
+        public double min = Double.MIN_VALUE;
+        public double max = Double.MAX_VALUE;
     }
 
     public boolean hasFunction(String fcn) {
